internal/create: add vcr-compatible `UniqueId` helper
This will replace calls to `sdkid.UniqueId` so that client token generation will be deterministic when go-vcr is enabled.
